Revert "database lifting (int -> float)"
This reverts commit 31377f40cdf2423bf3200e0593d6bb4e15631332.

+++ Wg-backend-api/Controllers/GameControllers/ResourcesController.cs
@@ -205,7 +205,7 @@ namespace Wg_backend_api.Controllers.GameControllers
                 .ToListAsync();
         }
 
-        private float GetCurrentResourceAmount(Nation nation, int resourceId)
+        private int GetCurrentResourceAmount(Nation nation, int resourceId)
         {
             return nation.Localisations
                 .SelectMany(l => l.LocalisationResources)
@@ -213,7 +213,7 @@ namespace Wg_backend_api.Controllers.GameControllers
                 .Sum(lr => lr.Amount);
         }
 
-        private float GetArmyMaintenanceExpenses(Nation nation, int resourceId)
+        private int GetArmyMaintenanceExpenses(Nation nation, int resourceId)
         {
             return nation.Armies
                 .SelectMany(a => a.Troops)
@@ -223,7 +223,7 @@ namespace Wg_backend_api.Controllers.GameControllers
                 .Sum();
         }
 
-        private float GetPopulationExpenses(Nation nation, int resourceId)
+        private int GetPopulationExpenses(Nation nation, int resourceId)
         {
             return nation.Localisations
                 .SelectMany(l => l.Populations)
@@ -233,7 +233,7 @@ namespace Wg_backend_api.Controllers.GameControllers
                 .Sum();
         }
 
-        private float GetPopulationProduction(Nation nation, int resourceId)
+        private int GetPopulationProduction(Nation nation, int resourceId)
         {
             return nation.Localisations
                 .SelectMany(l => l.Populations)
